Add files via upload
Attempt to write python code that pre-processes images of sunglasses before being put into Meshroom for photogrammetry. The idea was to reduce the negative impact of highly reflective surfaces by passing the image through a Canny edge detector to emphasize the boundaries of the physical object. Just doing the canny filter created thick edges that had jagged aliasing edges and created ambiguity for Meshroom, so I tried creating a complimentary filter of weighted to 90% the phase FFT image and 10% the Canny image. The reason I did not try using just the FFT phase was because the edges were not very pronounced. Ultimately this test did not result in better photogrammetric scans, at least for the highly reflective sunglasses frames I was scanning. Provided here are just one image sample. The whole dataset was ~200 images taken from all directions.